safari-600.7-branch/LayoutTests/platform/ios-sim/accessibility/tables-lists-expected.txt

item 1
list item 1
table item 1	table item 2	table item 3
landmark 1
This test makes sure that items in lists, tables and landmarks have different traits on them, because they're contained in different items.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS listItem.iphoneTraits != initialItem.iphoneTraits is true
PASS listItem.iphoneTraits != tableItem.iphoneTraits is true
FAIL listItem.iphoneTraits != landmarkItem.iphoneTraits should be true. Threw exception TypeError: undefined is not an object (evaluating 'landmarkItem.iphoneTraits')
PASS tableItem.iphoneTraits != initialItem.iphoneTraits is true
FAIL tableItem.iphoneTraits != landmarkItem.iphoneTraits should be true. Threw exception TypeError: undefined is not an object (evaluating 'landmarkItem.iphoneTraits')
FAIL landmarkItem.iphoneTraits != initialItem.iphoneTraits should be true. Threw exception TypeError: undefined is not an object (evaluating 'landmarkItem.iphoneTraits')
PASS successfullyParsed is true

TEST COMPLETE
